feat: split messages.ts into formValidation.ts and customerDashboard.ts
