+++ frontend/src/app/helper/messages/customerDashboard.ts
@@ -1,0 +1,2 @@
+export const FAILED_TO_FETCH_BOOKABLE_CLASSES =
+  "Failed to retrieve bookable classes.";
