+++ frontend/src/app/helper/utils/validationErrorUtils.ts
@@ -1,5 +1,5 @@
 import { ZodIssue } from "zod";
-import { GENERAL_ERROR_MESSAGE } from "./messages";
+import { GENERAL_ERROR_MESSAGE } from "../messages/formValidation";
 
 export function extractRegisterValidationErrors(
   validationErrors: ZodIssue[],
